[move] ファイルを特定の場所に配置
